+++ src/traits/i-icon/i-icon.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-unused-vars-experimental */
+
 /*!
  * V4Fire Client Core
  * https://github.com/V4Fire/Client
@@ -20,9 +22,11 @@ import { icons, iconsMap } from 'traits/i-icon/modules/icons';
 export default abstract class iIcon {
 	/**
 	 * Returns a link for the specified icon
+	 *
+	 * @param component
 	 * @param iconId
 	 */
-	static getIconLink(iconId: Nullable<string>): Promise<CanUndef<string>> {
+	static getIconLink: AddSelf<iIcon['getIconLink'], iBlock> = (component, iconId) => {
 		if (iconId == null) {
 			return SyncPromise.resolve(undefined);
 		}
@@ -49,7 +53,7 @@ export default abstract class iIcon {
 		}
 
 		return SyncPromise.resolve(`${location.pathname + q}#${icon.id}`);
-	}
+	};
 
 	/**
 	 * Updates `href` of the specified `use` element
@@ -58,7 +62,7 @@ export default abstract class iIcon {
 	 * @param el
 	 * @param [href]
 	 */
-	static updateIconHref<T extends iBlock>(component: T, el: SVGUseElement, href?: string): void {
+	static updateIconHref: AddSelf<iIcon['updateIconHref'], iBlock> = (component, el: SVGUseElement, href?) => {
 		const
 			$a = component.unsafe.async,
 			group = {group: el.getAttribute(ID_ATTRIBUTE) ?? undefined};
@@ -94,7 +98,7 @@ export default abstract class iIcon {
 				parent.removeChild(newEl);
 			} catch {}
 		}, group);
-	}
+	};
 
 	/**
 	 * Handles an error of the icon loading
@@ -103,15 +107,17 @@ export default abstract class iIcon {
 	 * @param el - link to the source `use` element
 	 * @param err
 	 */
-	static handleIconError<T extends iBlock & iIcon>(component: T, el: SVGUseElement, err: Error): void {
+	static handleIconError: AddSelf<iIcon['handleIconError'], iBlock & iIcon> = (component, el, err) => {
 		stderr(err);
 		component.updateIconHref(el);
-	}
+	};
 
 	/**
 	 * Link to iIcon.getIconLink
 	 */
-	abstract getIconLink: typeof iIcon.getIconLink;
+	getIconLink(iconId: Nullable<string>): Promise<CanUndef<string>> {
+		return <any>null;
+	}
 
 	/**
 	 * Updates `href` of the specified `use` element
@@ -119,7 +125,9 @@ export default abstract class iIcon {
 	 * @param el
 	 * @param [href]
 	 */
-	abstract updateIconHref(el: SVGUseElement, href?: string): void;
+	updateIconHref(el: SVGUseElement, href?: string): void {
+		return <any>null;
+	}
 
 	/**
 	 * Handles an error of the icon loading
@@ -127,5 +135,7 @@ export default abstract class iIcon {
 	 * @param el - link to the source `use` element
 	 * @param err
 	 */
-	abstract handleIconError(el: SVGUseElement, err: Error): void;
+	handleIconError(el: SVGUseElement, err: Error): void {
+		return <any>null;
+	}
 }